Clean up openai
