+++ mlflow/openai/autolog.py
@@ -1,6 +1,8 @@
 import functools
+import importlib.metadata
 import json
 import logging
+import warnings
 from typing import Any, AsyncIterator, Iterator
 
 from packaging.version import Version
@@ -10,7 +12,9 @@ from mlflow.entities import SpanType
 from mlflow.entities.span import LiveSpan
 from mlflow.entities.span_event import SpanEvent
 from mlflow.entities.span_status import SpanStatusCode
+from mlflow.exceptions import MlflowException
 from mlflow.ml_package_versions import _ML_PACKAGE_VERSIONS
+from mlflow.openai.constant import FLAVOR_NAME
 from mlflow.openai.utils.chat_schema import set_span_chat_attributes
 from mlflow.tracing.assessment import MlflowClient
 from mlflow.tracing.constant import (
@@ -24,12 +28,155 @@ from mlflow.tracing.utils import (
     end_client_span_or_trace,
     start_client_span_or_trace,
 )
+from mlflow.utils.annotations import experimental
+from mlflow.utils.autologging_utils import autologging_integration
 from mlflow.utils.autologging_utils.config import AutoLoggingConfig
+from mlflow.utils.autologging_utils.safety import safe_patch
 
 MIN_REQ_VERSION = Version(_ML_PACKAGE_VERSIONS["openai"]["autologging"]["minimum"])
 MAX_REQ_VERSION = Version(_ML_PACKAGE_VERSIONS["openai"]["autologging"]["maximum"])
 
 _logger = logging.getLogger(__name__)
+
+
+@experimental
+def autolog(
+    disable=False,
+    exclusive=False,
+    disable_for_unsupported_versions=False,
+    silent=False,
+    log_traces=True,
+    log_models=True,
+):
+    """
+    Enables (or disables) and configures autologging from OpenAI to MLflow.
+    Raises :py:class:`MlflowException <mlflow.exceptions.MlflowException>`
+    if the OpenAI version < 1.0.
+
+    Args:
+        disable: If ``True``, disables the OpenAI autologging integration. If ``False``,
+            enables the OpenAI autologging integration.
+        exclusive: If ``True``, autologged content is not logged to user-created fluent runs.
+            If ``False``, autologged content is logged to the active fluent run,
+            which may be user-created.
+        disable_for_unsupported_versions: If ``True``, disable autologging for versions of
+            OpenAI that have not been tested against this version of the MLflow
+            client or are incompatible.
+        silent: If ``True``, suppress all event logs and warnings from MLflow during OpenAI
+            autologging. If ``False``, show all events and warnings during OpenAI
+            autologging.
+        log_traces: If ``True``, traces are logged for OpenAI models. If ``False``, no traces are
+            collected during inference. Default to ``True``.
+        log_models: If ``True``, automatically create a LoggedModel when the model
+            used for inference is not already logged. The created LoggedModel contains no model
+            artifacts, but it will be used to associate all traces generated by the model. If
+            ``False``, no LoggedModel is created and the traces will not be associated with any
+            model. Default to ``True``.
+            .. Note:: Experimental: This argument may change or be removed in a future release
+            without warning.
+    """
+    if Version(importlib.metadata.version("openai")).major < 1:
+        raise MlflowException("OpenAI autologging is only supported for openai >= 1.0.0")
+
+    # This needs to be called before doing any safe-patching (otherwise safe-patch will be no-op).
+    # TODO: since this implementation is inconsistent, explore a universal way to solve the issue.
+    _autolog(
+        disable=disable,
+        exclusive=exclusive,
+        disable_for_unsupported_versions=disable_for_unsupported_versions,
+        silent=silent,
+        log_traces=log_traces,
+        log_models=log_models,
+    )
+
+    # Tracing OpenAI Agent SDK. This has to be done outside the function annotated with
+    # `@autologging_integration` because the function is not executed when `disable=True`.
+    try:
+        from mlflow.openai._agent_tracer import (
+            add_mlflow_trace_processor,
+            remove_mlflow_trace_processor,
+        )
+
+        if log_traces and not disable:
+            add_mlflow_trace_processor()
+        else:
+            remove_mlflow_trace_processor()
+    except ImportError:
+        pass
+
+
+# This is required by mlflow.autolog()
+autolog.integration_name = FLAVOR_NAME
+
+
+# NB: The @autologging_integration annotation must be applied here, and the callback injection
+# needs to happen outside the annotated function. This is because the annotated function is NOT
+# executed when disable=True is passed. This prevents us from removing our callback and patching
+# when autologging is turned off.
+@autologging_integration(FLAVOR_NAME)
+def _autolog(
+    disable=False,
+    exclusive=False,
+    disable_for_unsupported_versions=False,
+    silent=False,
+    log_traces=True,
+    log_models=True,
+):
+    from openai.resources.chat.completions import AsyncCompletions as AsyncChatCompletions
+    from openai.resources.chat.completions import Completions as ChatCompletions
+    from openai.resources.completions import AsyncCompletions, Completions
+    from openai.resources.embeddings import AsyncEmbeddings, Embeddings
+
+    for task in (ChatCompletions, Completions, Embeddings):
+        safe_patch(FLAVOR_NAME, task, "create", patched_call)
+
+    for task in (AsyncChatCompletions, AsyncCompletions, AsyncEmbeddings):
+        safe_patch(FLAVOR_NAME, task, "create", async_patched_call)
+
+    try:
+        from openai.resources.beta.chat.completions import AsyncCompletions, Completions
+    except ImportError:
+        pass
+    else:
+        safe_patch(FLAVOR_NAME, Completions, "parse", patched_call)
+        safe_patch(FLAVOR_NAME, AsyncCompletions, "parse", async_patched_call)
+
+    try:
+        from openai.resources.responses import AsyncResponses, Responses
+    except ImportError:
+        pass
+    else:
+        safe_patch(FLAVOR_NAME, Responses, "create", patched_call)
+        safe_patch(FLAVOR_NAME, AsyncResponses, "create", async_patched_call)
+
+    # Patch Swarm agent to generate traces
+    try:
+        from swarm import Swarm
+
+        warnings.warn(
+            "Autologging for OpenAI Swarm is deprecated and will be removed in a future release. "
+            "OpenAI Agent SDK is drop-in replacement for agent building and is supported by "
+            "MLflow autologging. Please refer to the OpenAI Agent SDK documentation "
+            "(https://github.com/openai/openai-agents-python) for more details.",
+            category=FutureWarning,
+            stacklevel=2,
+        )
+
+        safe_patch(
+            FLAVOR_NAME,
+            Swarm,
+            "get_chat_completion",
+            patched_agent_get_chat_completion,
+        )
+
+        safe_patch(
+            FLAVOR_NAME,
+            Swarm,
+            "run",
+            patched_swarm_run,
+        )
+    except ImportError:
+        pass
 
 
 def _get_span_type(task: type) -> str:
