+++ mlflow/openai/constant.py
@@ -1,0 +1,1 @@
+FLAVOR_NAME = "openai"
