+++ tests/openai/conftest.py
@@ -5,7 +5,7 @@ import mlflow
 
 from tests.helper_functions import start_mock_openai_server
 
-is_v1 = Version(mlflow.openai._get_openai_package_version()).major >= 1
+is_v1 = Version(mlflow.openai.model._get_openai_package_version()).major >= 1
 
 
 @pytest.fixture(scope="module", autouse=True)
